[REFACTOR] Context Provider
