+++ hooks/index.ts
@@ -1,0 +1,3 @@
+import { useCursorContext } from "./use-context/useContext";
+
+export { useCursorContext };
