+++ hooks/use-context/useContext.ts
@@ -1,0 +1,4 @@
+import { useContext } from "react";
+import { CursorContext } from "../../setup";
+
+export const useCursorContext = () => useContext(CursorContext);
