+++ setup/context-provider/contexts.types.ts
@@ -1,0 +1,15 @@
+interface CursorSettingsParams {
+  variant: string;
+  title: string | React.ReactNode;
+}
+
+interface CursorContextParams {
+  cursorSettings: CursorSettingsParams;
+  projectCursor: () => void;
+  defaultCursor: () => void;
+  otherProjectCursor: (title: string) => void;
+}
+
+interface ContextProviderProps {
+  children: React.ReactNode;
+}
