+++ setup/index.ts
@@ -1,3 +1,4 @@
 import ContextProvider from "./context-provider/ContextProvider";
+import { CursorContext } from "./context-provider/CursorContextProvider";
 
-export { ContextProvider };
+export { ContextProvider, CursorContext };
